requirements.txt
pymssql>=2.2.5
pyyaml>=6.0
msal>=1.20.0
psutil>=5.9.4
requests>=2.28.2